refactor(repo): split oversized Kotlin sources
