+++ ui/src/desktopTest/kotlin/dev/obiente/nextcloudnative/app/JvmSupportIntakeTestSupport.kt
@@ -1,0 +1,4 @@
+package dev.obiente.nextcloudnative.app
+
+internal fun JvmSupportIntake.submittedRecordId(): String =
+    (states().value as SupportDiagnosticsSubmissionState.Submitted).reports.single().recordId
